refactor tests for options django app
Conflicts:
	apps/options/testing/factories.py
